Move operations to separate module
